float

+++ database/jeux_olympiques.sql
@@ -48,8 +48,8 @@ CREATE TABLE `sport` (
  `id_sport` TINYINT UNSIGNED PRIMARY KEY AUTO_INCREMENT,
  `Nom` varchar(255) NOT NULL,
  `Image` varchar(255) NOT NULL,
- `latitude` float(255) NOT NULL,
- `longitude` float(255) NOT NULL
+ `latitude` float(25) NOT NULL,
+ `longitude` float(25) NOT NULL
 
 
 );
